github: change how UA tokens are used
With recent changes to snapcraft [1], --ua-token doesn't work. A change
is in the works to use Pro LXD containers which will adopt the Pro
status from the host machine to create their environment.

Because this change will (probably) require SNAPCRAFT_UA_TOKEN to be set
and using that variable works for earlier base snaps, we might as well
adopt using it for all snaps we build.

[1] https://github.com/canonical/snapcraft/issues/4874

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>