gadget.yaml: add a kernel commandline allow list
This is to enable adding some options, pertinent for my use-case,
specifically pulling the options from a lovely gist:
https://gist.github.com/desrod/cc857bbadc3d8ac34770e2624a8fc76e

Some earlier changes are reverted, specifically partition information;
this is because I have a NAS deployed currently using an older
gadget.yaml partition structure, and modifying the volume layout is an
unsupportable change. We would need to completely reimage the machine.
Note that this IS planned, but too much testing is happening at the
moment.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>